refactor the bigquery directory layout
