+++ tensorflow_io/core/BUILD
@@ -773,7 +773,7 @@ cc_binary(
     linkshared = 1,
     deps = [
         "//tensorflow_io/core:arrow_ops",
-        "//tensorflow_io/bigquery:bigquery_ops",
+        "//tensorflow_io/core:bigquery_ops",
         "//tensorflow_io/core:audio_video_ops",
         "//tensorflow_io/core:avro_ops",
         "//tensorflow_io/core:orc_ops",
@@ -864,4 +864,53 @@ cc_binary(
     deps = [
         "//tensorflow_io/core:ffmpeg_2.8_ops",
     ],
+)
+
+# Have to put test op into prod binary because of https://github.com/grpc/grpc/issues/20034
+# Otherwise whenever 2 dynamically built libraries that are using grpc loaded
+# using _load_library, only first one will work.
+cc_library(
+    name = "bigquery_ops",
+    srcs = [
+        "kernels/bigquery/bigquery_dataset_op.cc",
+        "kernels/bigquery/bigquery_kernels.cc",
+        "kernels/tests/bigquery_test_client_op.cc",
+        "ops/bigquery_ops.cc",
+        "ops/bigquery_test_ops.cc",
+    ],
+    copts = tf_io_copts(),
+    linkstatic = True,
+    deps = [
+        ":bigquery_lib_cc",
+        "@arrow",
+        "@avro",
+        "@com_github_grpc_grpc//:grpc++",
+        "@com_google_googleapis//google/cloud/bigquery/storage/v1beta1:storage_cc_grpc",
+        "@local_config_tf//:libtensorflow_framework",
+        "@local_config_tf//:tf_header_lib",
+    ],
+    alwayslink = 1,
+)
+
+# A library for use in the bigquery kernels.
+cc_library(
+    name = "bigquery_lib_cc",
+    srcs = ["kernels/bigquery/bigquery_lib.cc"],
+    hdrs = ["kernels/bigquery/bigquery_lib.h"],
+    copts = tf_io_copts(),
+    linkstatic = True,
+    deps = [
+        "//tensorflow_io/core:arrow_util",
+        "@arrow",
+        "@avro",
+        "@com_github_grpc_grpc//:grpc++",
+        "@com_google_absl//absl/algorithm",
+        "@com_google_absl//absl/container:fixed_array",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/types:variant",
+        "@com_google_googleapis//google/cloud/bigquery/storage/v1beta1:storage_cc_grpc",
+        "@local_config_tf//:libtensorflow_framework",
+        "@local_config_tf//:tf_header_lib",
+    ],
+    alwayslink = 1,
 )
